Create create.blade.php
This is an attempt at the view that will allow the user to add a new site to the directory. It has not been tested just yet. Please check it over and give me any comments you might have. I am still working on making it function correctly and I am thinking the other fields just self populate???